requirements.txt
Django==3.1.4
ipython==7.19.0
jedi==0.17.2
django-extensions==3.1.0

# Python service calls
#requests==1.2.3

# System monitoring and Performance
#raven==2.0.3

coverage==5.3.1
pep8==1.7.1
pyflakes==2.2.0
pytest==6.2.1

# For API
djangorestframework==3.12.2

# For Gamestate
redis==3.5.3
mysqlclient==2.0.3

